Persists last indexed height during sync
Adds checkpointing to the syncing datasource. This allows the
indexer to resume from the last indexed height in case of a
restart. A new `CheckpointStore` trait is introduced and used to
persist the last indexed height during live syncing.

Also introduces a utility function to determine the maximum height
within an `Updates` payload, which is used for checkpointing.